Seungpyo

+++ controllers/choice.controller.js
@@ -7,7 +7,7 @@ class ChoiceController {
     try {
       const { userKey } = res.locals.user;
       if (userKey == 0) {
-        res.status(400).send({ message: "로그인이 필요합니다." });
+        return res.status(400).send({ message: "로그인이 필요합니다." });
       }
       const { title, choice1Name, choice2Name, endTime } = req.body;
 
@@ -29,6 +29,7 @@ class ChoiceController {
       next(err);
     }
   };
+
 
   //모든 게시글 조회(1~10)
   allchoice = async (req, res, next) => {
@@ -54,12 +55,13 @@ class ChoiceController {
     }
   };
 
+
   //마이페이지 게시글 조회
   mychoice = async (req, res, next) => {
     try {
       const { userKey } = res.locals.user;
       if (userKey == 0) {
-        res.status(400).send({ message: "로그인이 필요합니다." });
+        return res.status(400).send({ message: "로그인이 필요합니다." });
       }
       const mychoice = await this.choiceService.findMychoice(userKey);
       res.status(200).json({ data: mychoice });
@@ -73,7 +75,9 @@ class ChoiceController {
       const { userKey } = res.locals.user;
       const { choiceId } = req.params;
       if (!choiceId) throw new Error("없는 게시글 입니다.");
-
+      if (userKey == 0) {
+        return res.status(400).send({ message: "권한이 없습니다." });
+      }
       const deletechoice = await this.choiceService.deletechoice(
         userKey,
         choiceId
@@ -95,7 +99,7 @@ class ChoiceController {
     try {
       const { userKey } = res.locals.user;
       if (userKey == 0) {
-        res.status(400).send({ message: "로그인이 필요합니다." });
+        return res.status(400).send({ message: "로그인이 필요합니다." });
       }
       const { choiceId } = req.params;
       const { choiceNum } = req.body;
